requirements.txt
fastapi==0.61.1
uvicorn==0.12.2
font-source-sans-pro==0.0.1
inky==1.0.1
numpy==1.19.2
Pillow==8.0.0
RPi.GPIO==0.7.0
six==1.15.0
smbus2==0.3.0
spidev==3.5